+++ wp-content/plugins/menu-icons/CHANGELOG.md
@@ -1,7 +1,12 @@
 
- ### v0.12.2 - 2020-04-27 
+ ### v0.12.4 - 2020-07-13 
  **Changes:** 
- * Add Menu Icons extension
+ * Fix Font Awesome not loading
+ 
+ ### v0.12.3 - 2020-07-13 
+ **Changes:** 
+ * Fixed Menu Icons in Block Editor not working
+* Fixed CWP links.
  
  ### v0.12.2 - 2019-11-15 
  **Changes:**